rename lpw -> lpw_stable_diffusion
